pyproject.toml
[tool.poetry]
name = "docx-report"
version = "0.1.2"
description = "A wrapper for the python-docx library that simplifies creating Word documents in Python"
authors = ["Chris Hacker <49451910+christopher-hacker@users.noreply.github.com>"]
readme = "README.rst"
packages = [
    { include = "*", from = "docx_report" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-docx = "^1.0.1"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
pyjanitor = "^0.26.0"
sphinx-click = "^5.0.1"
recommonmark = "^0.7.1"
pydata-sphinx-theme = "^0.14.1"
sphinx-autodoc-typehints = "^1.24.0"
pytz = "^2023.3.post1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pylint = "^3.0.2"
pytest = "^7.4.2"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.5.0"
sphinx = "^7.2.6"
sphinx-autodoc-typehints = "^1.24.0"

[tool.pytest.ini_options]
addopts = "--cov=docx_report --cov-report term --cov-report html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
